requirements.txt
flask>=3.0
requests>=2.31
pandas>=2.0
numpy>=1.26
pytz>=2024.1
pyyaml>=6.0
schedule>=1.2
